GCRApps/dodo/getpage/requirements.txt
Flask==2.0.1
dataclasses==0.6
requests==2.26.0
beautifulsoup4==4.8.0
protobuf==3.19.0
python_dateutil==2.8.2
google-cloud-bigquery==2.26.0
gunicorn==20.1.0
google-cloud-pubsub==2.9.0
google-cloud-spanner==3.15.1